+++ CHANGELOG.md
@@ -10,6 +10,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Security
 - Docker file to update apk packages on build @philip-r-beckwith [#1309](https://github.com/newrelic/nri-kubernetes/pull/1309)
 
+## v3.47.0 - 2025-10-20
+
+### 🚀 Enhancements
+- Add metrics for ResourceQuota @NRhzhao [#1302](https://github.com/newrelic/nri-kubernetes/pull/1302)
+
+### ⛓️ Dependencies
+- Updated go to v1.25.3
+
+## v3.46.0 - 2025-10-13
+
+### 🚀 Enhancements
+- Add v1.34 support and drop support for v1.29 @TmNguyen12 [#1300](https://github.com/newrelic/nri-kubernetes/pull/1300)
+
+### ⛓️ Dependencies
+- Updated golang.org/x/crypto to v0.42.0
+- Updated alpine to v3.22.2
+- Updated kubernetes packages to v0.34.1
+
 ## v3.45.4 - 2025-10-06
 
 ### 🐞 Bug fixes